+++ libs/shared/ui/src/lib/components/button-primitive/button-primitive.spec.tsx
@@ -1,0 +1,9 @@
+import { renderWithProviders } from '@qovery/shared/util-tests'
+import { ButtonPrimitive } from './button-primitive'
+
+describe('ButtonPrimitive', () => {
+  it('should match snapshot', () => {
+    const { baseElement } = renderWithProviders(<ButtonPrimitive type="button">Foobar</ButtonPrimitive>)
+    expect(baseElement).toMatchSnapshot()
+  })
+})
