+++ libs/shared/ui/src/lib/components/button/button.spec.tsx
@@ -2,8 +2,12 @@ import { renderWithProviders } from '@qovery/shared/util-tests'
 import { Button } from './button'
 
 describe('Button', () => {
-  it('should match snapshot', () => {
-    const { baseElement } = renderWithProviders(<Button type="button">Foobar</Button>)
+  it('should match snapshot with loading', () => {
+    const { baseElement } = renderWithProviders(
+      <Button type="button" loading>
+        Foobar
+      </Button>
+    )
     expect(baseElement).toMatchSnapshot()
   })
 })
